refactor: detail, form 폴더 분리
