+++ src/components/vote/detail/ParticipationChart.tsx
